+++ packages/dask-task-models-library/src/dask_task_models_library/models.py
@@ -1,0 +1,4 @@
+from typing import TypeAlias
+
+DaskJobID: TypeAlias = str
+DaskResources: TypeAlias = dict[str, int | float]
