+++ services/director-v2/src/simcore_service_director_v2/modules/dask_client.py
@@ -37,6 +37,7 @@ from dask_task_models_library.container_tasks.protocol import (
     LogFileUploadURL,
     TaskOwner,
 )
+from dask_task_models_library.container_tasks.utils import generate_dask_job_id
 from dask_task_models_library.resource_constraints import (
     create_ec2_resource_constraint_key,
 )
@@ -310,7 +311,7 @@ class DaskClient:
 
         list_of_node_id_to_job_id: list[PublishedComputationTask] = []
         for node_id, node_image in tasks.items():
-            job_id = dask_utils.generate_dask_job_id(
+            job_id = generate_dask_job_id(
                 service_key=node_image.name,
                 service_version=node_image.tag,
                 user_id=user_id,
